+++ changelog.md
@@ -12,6 +12,8 @@ While all _Major_ releases should be reviewed, our only _large_ releases are **v
 #### 14.5.0 [August 2022]
 - **[fix]** - possible runtime error in setTag method
 - **[change]** - make #Honorific always a #Person #951
+- **[new]** - manually change conjugations/inflections from plugin #949
+- **[new]** - `.adjectives().conjugate()` method
 
 #### 14.4.5 [August 2022]
 - **[fix]** - fix logic for greedy-negative matches - #936